+++ client/package.json
@@ -12,6 +12,7 @@
     "babel-eslint": "^10.1.0",
     "jwt-decode": "^3.1.2",
     "moment": "^2.27.0",
+    "node-sass": "^7.0.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-file-base64": "^1.0.3",
