+++ client/src/components/Container/classes.module.scss
@@ -1,0 +1,3 @@
+.pageContainer {
+  padding: 20px 50px;
+};
