+++ client/src/components/index.js
@@ -1,5 +1,7 @@
 import Navbar from './Navbar/Navbar';
+import { PageContainer } from './Container';
 
 export {
   Navbar,
+  PageContainer,
 };
