+++ client/src/index.css
@@ -4,3 +4,7 @@
     /* font-family: 'Noto Sans JP' !important; */
     font-family: 'Heebo', sans-serif !important;
 }
+
+.ant-layout-content {
+    min-height: 85vh;
+}
