+++ client/src/pages/home/components/index.js
@@ -1,0 +1,2 @@
+export { HomePageSlider } from './Slider.js';
+export { Benefits } from './Benefits';
