+++ client/src/pages/index.js
@@ -4,6 +4,7 @@ import AuthPage from './AuthPage';
 import CreatorOrTag from './CreatorOrTag';
 import PageTimeshit from './timeshit';
 import PageProfessionals from './professionals';
+import HomePage from './home';
 
 export {
   PostDetails,
@@ -12,4 +13,5 @@ export {
   CreatorOrTag,
   PageTimeshit,
   PageProfessionals,
+  HomePage,
 };
